GSC Portfolio Card states (#65)
* Renamed to GSCButtons and added toast notfications to join and leave transactions

* Create useGSCStatus.ts

Hook to reconcile which state the connected account is in. Created an enum that is used to track the state which is used in various components.

* option parameter for hooks

* Delete JoinGSCButton.tsx

* Create useIsMemberKicked.ts

checks if an account has previously been kicked

* Cleanup and responsiveness updates for portfolio card

* import path update and misc

* Create GSCHistory.tsx

Component for displaying GSC history and informationals depending on gsc status

* yarn update

* addressed comments

changed account to match params of other hooks. cleaned up nested if statements in useGSCstatus

* added back if statment for current state and using formatEther instead of formatUnits